scraper/email-verification-service/requirements.txt
flask==2.3.3
requests==2.31.0
msal==1.24.0
python-dotenv==1.0.0